Switched to linear comparison of chain clusters

+++ src/seed_clusterer.hpp
@@ -36,7 +36,7 @@ class SnarlSeedClusterer {
                            snarls_to_node_t& snarls_to_node, 
                            node_to_seed_t& node_to_seed);
 
-        tuple<hash_set<size_t>, size_t, size_t> 
+        tuple<hash_set<size_t>, int64_t, int64_t> 
              get_clusters_node(const vector<pos_t>& seeds, 
                              structures::UnionFind& union_find_clusters,
                              hash_map<size_t, pair<int64_t, int64_t>>& cluster_dists,
@@ -45,7 +45,7 @@ class SnarlSeedClusterer {
                              DistanceIndex& dist_index, id_t root, 
                              int64_t node_length);
 
-        tuple<hash_set<size_t>, size_t, size_t> get_clusters_chain(
+        tuple<hash_set<size_t>, int64_t, int64_t> get_clusters_chain(
                              const vector<pos_t>& seeds,
                              structures::UnionFind& union_find_clusters,
                              hash_map<size_t, pair<int64_t, int64_t>>& cluster_dists,
@@ -55,7 +55,7 @@ class SnarlSeedClusterer {
                              size_t distance_limit, const SnarlManager& snarl_manager,
                              DistanceIndex& dist_index, const Chain* root);
 
-        tuple<hash_set<size_t>, size_t, size_t> get_clusters_snarl(
+        tuple<hash_set<size_t>, int64_t, int64_t> get_clusters_snarl(
                              const vector<pos_t>& seeds,
                              structures::UnionFind& union_find_clusters,
                              hash_map<size_t, pair<int64_t, int64_t>>& cluster_dists,
